refactor: move hashing and add simpler labeling logic
