+++ src/orcabridge/mapper.py
@@ -8,7 +8,7 @@ from .utils.stream_utils import (
     batch_tag,
     batch_packet,
 )
-from .utils.hash import function_content_hash, stable_hash
+from .hashing import function_content_hash, stable_hash
 from .types import Tag, Packet
 from typing import Iterator, Tuple
